Lab 1/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
dbis-er-diagram==1.1.2
decorator==5.2.1
executing==2.2.1
google-auth==2.41.1
google-genai==1.42.0
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Levenshtein==0.27.1
matplotlib-inline==0.1.7
networkx==3.5
numpy==2.3.3
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
traitlets==5.14.3
typeguard==3.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
